Optimize image files and simplify sprite rendering

+++ src/main/java/com/wildfire/render/WildfireModelRenderer.java
@@ -26,7 +26,11 @@ import com.wildfire.main.uvs.UVLayout;
 import com.wildfire.main.uvs.UVQuad;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
-import org.joml.*;
+import org.joml.Matrix3f;
+import org.joml.Matrix4f;
+import org.joml.Vector3f;
+import org.joml.Vector3fc;
+import org.joml.Vector4f;
 
 @Environment(EnvType.CLIENT)
 public final class WildfireModelRenderer {
@@ -73,15 +77,15 @@ public final class WildfireModelRenderer {
             this.posX1 = x;
             this.posY1 = y;
             this.posZ1 = z;
-            this.posX2 = x + (float) dx;
-            this.posY2 = y + (float) dy;
-            this.posZ2 = z + (float) dz;
+            this.posX2 = x + dx;
+            this.posY2 = y + dy;
+            this.posZ2 = z + dz;
             this.quads = new TexturedQuad[quads];
             this.dynamicUvLayouts = dynamicUvLayouts;
 
-            float f = x + (float) dx;
-            float f1 = y + (float) dy;
-            float f2 = z + (float) dz;
+            float f = x + dx;
+            float f1 = y + dy;
+            float f2 = z + dz;
             x = x - delta;
             y = y - delta;
             z = z - delta;
